% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{alert_generate}
\alias{alert_generate}
\title{Alert message generation function}
\usage{
alert_generate(text, .envir = parent.frame())
}
\arguments{
\item{text}{Generated text output.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
Alert message generation function
}
