% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch.process}
\alias{batch.process}
\title{Batch read in backscatter export files.}
\usage{
batch.process(directory, return_object = T, write_file = F, save_directory)
}
\arguments{
\item{directory}{File path where files are saved.}

\item{return_object}{Return object (tibble). (Default: \code{T})}

\item{write_file}{Write file. (Default: \code{F})}

\item{save_directory}{File path where processed files will be saved.}
}
\description{
Batch read in backscatter export files.
}
