% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch.read}
\alias{batch.read}
\title{Batch read in backscatter export files.}
\usage{
batch.read(directory)
}
\arguments{
\item{directory}{File path where files are saved.}
}
\description{
Batch read in backscatter export files.
}
