% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{degrees}
\alias{degrees}
\title{Convert from degrees to radians}
\usage{
degrees(coordinate)
}
\arguments{
\item{coordinate}{Longitude/Latitude (radians).}
}
\description{
Convert from degrees to radians
}
