% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.nasc}
\alias{integrate.nasc}
\title{Vertically integrate backscatter to calculate NASC}
\usage{
integrate.nasc(data)
}
\arguments{
\item{data}{Export tibble.}
}
\description{
Vertically integrate backscatter to calculate NASC
}
