% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{line.distance}
\alias{line.distance}
\title{Calculate distance between two points}
\usage{
line.distance(Longitude, Latitude)
}
\arguments{
\item{Longitude}{Longitude (degrees).}

\item{Latitude}{Latitude (degrees).}
}
\description{
Calculate distance between two points
}
