% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.nasc}
\alias{plot.nasc}
\title{Stratified mean NASC plot}
\usage{
\method{plot}{nasc}(
  data,
  frequency = c(38, 70, 120, 200),
  month_pad = 2,
  season_definition = "meteorological",
  point.size = 3.25,
  line.width = 2,
  text.main.size = 15.25,
  text.axis.size = 14.25,
  text.legend.size = 14.5,
  text.label.size = 5.5,
  background_alpha = 0.4,
  common_axis = F
)
}
\arguments{
\item{data}{Export tibble.}

\item{frequency}{Numeric frequencies (kHz) that will be plotted.}

\item{month_pad}{Number of months to pad along the x-axis.}

\item{season_definition}{Season definition (meteorological vs astronomical).}

\item{point.size}{Plot point size.}

\item{line.width}{Plot line width.}

\item{text.main.size}{Main text size.}

\item{text.axis.size}{Axis text size.}

\item{text.legend.size}{Legend text size.}

\item{text.label.size}{Label text size.}

\item{background_alpha}{Season color transparency.}

\item{common_axis}{Fixed versus facetted scales.}
}
\description{
Stratified mean NASC plot
}
