% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{radians}
\alias{radians}
\title{Convert from degrees to radians}
\usage{
radians(coordinate)
}
\arguments{
\item{coordinate}{Longitude/Latitude (degrees).}
}
\description{
Convert from degrees to radians
}
