% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{read.export}
\alias{read.export}
\title{Read in backscatter export file}
\usage{
read.export(file, write_file = F, save_directory)
}
\arguments{
\item{file}{Export file (.csv).}

\item{write_file}{Write file. (Default: \code{F})}

\item{save_directory}{File path where processed files will be saved.}
}
\description{
Read in backscatter export file
}
