% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_interval}
\alias{seasonal_interval}
\title{Define seasonal interval based on dates}
\usage{
seasonal_interval(data, month_pad = 2, season_definition = "meteorological")
}
\arguments{
\item{data}{Export tibble.}

\item{month_pad}{Number of months to pad along the x-axis.}

\item{season_definition}{Season definition (meteorological vs astronomical).}
}
\description{
Define seasonal interval based on dates
}
