% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stratified_palette}
\alias{stratified_palette}
\title{Seasonal color palette}
\usage{
stratified_palette(starting_season)
}
\arguments{
\item{starting_season}{The first season that will be plotted (left-to-right)
along x-axis of plot.}
}
\description{
Seasonal color palette
}
