% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{survey.distance}
\alias{survey.distance}
\title{Wrapper function that calculates the along-transect distance}
\usage{
survey.distance(data)
}
\arguments{
\item{data}{Export tibble.}
}
\description{
Wrapper function that calculates the along-transect distance
}
