% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{transect.distance}
\alias{transect.distance}
\title{Calculate along-track/transect distance for each transect}
\usage{
transect.distance(Interval, Longitude, Latitude)
}
\arguments{
\item{Interval}{Interval (along-transect interval number).}

\item{Longitude}{Longitude (degrees).}

\item{Latitude}{Latitude (degrees).}
}
\description{
Calculate along-track/transect distance for each transect
}
