% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statified.mean}
\alias{statified.mean}
\title{Calculate the stratified mean (based on transect length)}
\usage{
statified.mean(data, ci_pct = 0.95)
}
\arguments{
\item{data}{Export tibble.}

\item{ci_pct}{Confidence interval percentile. (Default = 0.95)}
}
\description{
Calculate the stratified mean (based on transect length)
}
